clip-image-search/backend/app/requirements.txt
fastapi
uvicorn[standard]
pymongo
pillow
numpy
torch
torchvision
transformers
python-multipart
tqdm
